Take a BlockGetter where a Level has to fit
26.2 made BlockAndTintGetter a client-only view onto a chunk being meshed, and
Level no longer implements it. Everything that only reads block states and
block entities takes BlockGetter, the supertype the two share, so the same
helpers serve the world and the model pass; only the code that needs a meshing
view's tint and light keeps the narrower type.

+++ src/main/java/com/simibubi/create/compat/jei/DoubleItemIcon.java
@@ -42,17 +42,17 @@ public class DoubleItemIcon implements IDrawable {
 
 		RenderSystem.enableDepthTest();
 		matrixStack.pushMatrix();
-		matrixStack.translate(xOffset, yOffset);
+		matrixStack.translate((float) (xOffset), (float) (yOffset));
 
 		matrixStack.pushMatrix();
-		matrixStack.translate(1, 1);
+		matrixStack.translate((float) (1), (float) (1));
 		GuiGameElement.of(primaryStack)
 			.submit(graphics);
 		matrixStack.popMatrix();
 
 		matrixStack.pushMatrix();
-		matrixStack.translate(10, 10);
-		matrixStack.scale(.5f, .5f);
+		matrixStack.translate((float) (10), (float) (10));
+		matrixStack.scale((float) (.5f), (float) (.5f));
 		GuiGameElement.of(secondaryStack)
 			.submit(graphics);
 		matrixStack.popMatrix();
