+++ src/main/java/com/simibubi/create/compat/jei/category/MechanicalCraftingCategory.java
@@ -115,8 +115,8 @@ public class MechanicalCraftingCategory extends CreateRecipeCategory<CraftingRec
 					.isEmpty())
 					continue;
 				matrixStack.pushMatrix();
-				matrixStack.translate(col * 19 * scale, row * 19 * scale);
-				matrixStack.scale(scale, scale);
+				matrixStack.translate((float) (col * 19 * scale), (float) (row * 19 * scale));
+				matrixStack.scale((float) (scale), (float) (scale));
 				AllGuiTextures.JEI_SLOT.render(graphics, 0, 0);
 				matrixStack.popMatrix();
 			}
@@ -128,7 +128,7 @@ public class MechanicalCraftingCategory extends CreateRecipeCategory<CraftingRec
 		crafter.draw(graphics, 129, 25);
 
 		matrixStack.pushMatrix();
-		matrixStack.translate(0, 0);
+		matrixStack.translate((float) (0), (float) (0));
 
 		int amount = 0;
 		for (Ingredient ingredient : recipe.getIngredients()) {
@@ -156,7 +156,7 @@ public class MechanicalCraftingCategory extends CreateRecipeCategory<CraftingRec
 			Matrix3x2fStack matrixStack = graphics.pose();
 			matrixStack.pushMatrix();
 			float scale = getScale(recipe);
-			matrixStack.scale(scale, scale);
+			matrixStack.scale((float) (scale), (float) (scale));
 
 			if (ingredient != null) {
 				Matrix4fStack modelViewStack = RenderSystem.getModelViewStack();
