+++ src/main/java/com/simibubi/create/compat/jei/category/animations/AnimatedMixer.java
@@ -16,7 +16,7 @@ public class AnimatedMixer extends AnimatedKinetics {
 		viewRotation(-15.5f, 22.5f);
 		Matrix3x2fStack matrixStack = graphics.pose();
 		matrixStack.pushMatrix();
-		matrixStack.translate(xOffset, yOffset);
+		matrixStack.translate((float) (xOffset), (float) (yOffset));
 		int scale = 23;
 
 		blockElement(cogwheel())
