+++ src/main/java/com/simibubi/create/content/equipment/zapper/ZapperScreen.java
@@ -128,10 +128,10 @@ public abstract class ZapperScreen extends AbstractSimiScreen {
 	protected void renderBlock(GuiGraphicsExtractor graphics, int x, int y) {
 		Matrix3x2fStack ms = graphics.pose();
 		ms.pushMatrix();
-		ms.translate(x + 32, y + 42);
+		ms.translate((float) (x + 32), (float) (y + 42));
 		ms.mulPose(Axis.XP.rotationDegrees(-25f));
 		ms.mulPose(Axis.YP.rotationDegrees(-45f));
-		ms.scale(20, 20);
+		ms.scale((float) (20), (float) (20));
 
 		BlockState state = zapper.getOrDefault(AllDataComponents.SHAPER_BLOCK_USED, Blocks.AIR.defaultBlockState());;
